Initial methods for restructuration
