+++ contract/src/lib.rs
@@ -1,8 +1,9 @@
 use near_sdk::borsh::{self, BorshDeserialize, BorshSerialize};
 use near_sdk::{near_bindgen, env, setup_alloc, AccountId};
-use near_sdk::serde::{Serialize, Deserialize};
-use near_sdk::collections::UnorderedMap;
-use std::collections::HashSet;
+// use near_sdk::serde::{Serialize, Deserialize};
+use near_sdk::collections::{UnorderedMap, UnorderedSet};
+// use std::collections::HashSet;
+// use std::vec::Vec;
 use std::convert::From;
 
 setup_alloc!();
@@ -11,18 +12,23 @@ setup_alloc!();
 #[derive(BorshDeserialize, BorshSerialize)]
 pub struct GuildsPlatform {
     guilds: UnorderedMap<String, UpgradableGuild>,
-    users: UnorderedMap<AccountId, HashSet<String>>
+    users: UnorderedMap<AccountId, UnorderedSet<String>>
 }
 
-//Initializing the contract
 impl Default for GuildsPlatform {
     fn default() -> Self {
-      Self {
+        Self {
         guilds: UnorderedMap::new(b"g".to_vec()),
         users: UnorderedMap::new(b"u".to_vec())
-      }
+        }
     }
-  }
+}
+
+#[derive(BorshSerialize, BorshDeserialize)]
+pub struct Guild {
+    pub slug: String,
+    pub members: UnorderedSet<AccountId>,
+}
 
 /// Default trait for Guild
 impl Default for Guild {
@@ -31,17 +37,46 @@ impl Default for Guild {
     }
 }
 
-/// Default trait for UpgradableGuild
-impl Default for UpgradableGuild {
-    fn default() -> Self {
-        UpgradableGuild::CurrentVersion(new_guild())
+pub fn new_guild() -> Guild {
+    env::log(format!("Emplty guild created").as_bytes());
+    Guild {
+        slug: String::from(""),
+        members: UnorderedSet::new(b"n".to_vec()),
     }
 }
 
+#[near_bindgen]
+impl GuildsPlatform {
+    pub fn create_guild(&mut self, slug: String){
+        //Creates a guild based on the current version
+        let guild = UpgradableGuild::CurrentVersion(Guild {
+            slug: String::from(&slug),
+            members: UnorderedSet::new(slug.as_bytes().to_vec()),
+        });
+
+        self.guilds.insert(&slug, &guild);
+
+        env::log(format!("Saving guild '{}'", &slug,).as_bytes());
+    }
+
+    // pub fn get_guild_info(&self, slug: String) -> Guild {
+    //     let guild = self.guilds.get(&slug).unwrap_or_default().into();
+
+    //     return guild
+    // }
+
+    pub fn guild_exists(&self, slug: &String) ->bool {
+        match self.guilds.get(&slug) {
+            Some(_value) => {
+                true
+            },
+            None => false
+        }
+    }
+}
 
 //Implementation for upgradable contract
-#[derive(Serialize, Deserialize, BorshSerialize, BorshDeserialize, Debug)]
-#[serde(crate = "near_sdk::serde")]
+#[derive(BorshSerialize, BorshDeserialize)]
 pub enum UpgradableGuild {
     CurrentVersion(Guild),
 }
@@ -53,221 +88,3 @@ impl From<UpgradableGuild> for Guild {
         }
     }
 }
-
-#[derive(Serialize, Deserialize, BorshSerialize, BorshDeserialize, Debug)]
-#[serde(crate = "near_sdk::serde")]
-pub struct Guild {
-    pub slug: String,
-    pub title: String,
-    pub oneliner: String,
-    pub website: String,
-    pub app: String,
-    pub whitepaper: String,
-    pub twitter: String,
-    pub telegram: String,
-    pub discord: String,
-    pub medium: String,
-    pub github: String,
-    pub ticker: String,
-    pub logo: String,
-    pub contract_str: String,
-    pub youtube: String,
-    pub members: HashSet<AccountId>,
-}
-
-#[derive(BorshSerialize, BorshDeserialize)]
-pub enum GuildStatus {
-    Launched,
-    NotLaunched,
-}
-
-#[near_bindgen]
-impl GuildsPlatform {
-    pub fn create_guild(
-        &mut self, 
-        slug: String,
-        title: String,
-        oneliner: String,
-        website: String,
-        app: String,
-        whitepaper: String,
-        twitter: String,
-        telegram: String,
-        discord: String,
-        medium: String,
-        github: String,
-        ticker: String,
-        logo: String,
-        youtube: String,
-        contract_str: String,
-    ){
-        let guild = UpgradableGuild::CurrentVersion(Guild {
-            slug: String::from(&slug),
-            title: String::from(&title),
-            oneliner: String::from(&oneliner),
-            website: String::from(&website),
-            app: String::from(&app),
-            whitepaper: String::from(&whitepaper),
-            twitter: String::from(&twitter),
-            telegram: String::from(&telegram),
-            discord: String::from(&discord),
-            medium: String::from(&medium),
-            github: String::from(&github),
-            ticker: String::from(&ticker),
-            logo: String::from(&logo),
-            contract_str: String::from(&contract_str),
-            youtube: String::from(&youtube),
-            members: HashSet::new(),
-        });
-
-        self.guilds.insert(&slug, &guild);
-
-        env::log(format!("Saving guild '{}'", &slug,).as_bytes());
-    }
-
-    pub fn get_guild_info(&self, slug: String) -> Guild {
-        let guild: Guild = self.guilds.get(&slug).unwrap_or_default().into();
-
-        return guild
-    }
-
-    pub fn join_guild(&mut self, slug: String) {
-
-        //TO DO: Control over joining? Can anyone join?
-        //TO DO: On boarding of people with no account.
-        let mut guild: Guild = self.guilds.get(&slug).unwrap_or_default().into();
-        let account_to_insert = env::predecessor_account_id();
-        
-        if guild.members.get(&account_to_insert).is_none() {
-            guild.members.insert(account_to_insert);
-            self.guilds.insert(&slug,&UpgradableGuild::CurrentVersion(guild));
-            
-            //PROPOSAL: User methods
-            //We check if the register of the user already exists. 
-            //If it does, we insert the new guild. 
-            //if it doesn't, we create the registry and add that guild to the list.
-            let mut user_list: HashSet<AccountId>;
-            match self.check_if_user(){
-                true => {
-                    user_list = self.users.get(&env::predecessor_account_id()).unwrap_or_default();
-                },
-                false => {
-                    user_list = HashSet::new();
-                }
-            };
-
-            user_list.insert(slug.to_string());
-            self.users.insert(&env::predecessor_account_id(),&user_list);
-            //PROPOSAL ENDS
-
-            env::log(format!("'{}' just joined '{}'!", env::predecessor_account_id(), &slug,).as_bytes());
-        }
-        else{
-            env::log(format!("'{}' is already a member of '{}'!", account_to_insert, &slug,).as_bytes());
-        }
-        //TO DO: Recieve an NFT to confirm joining the guild?
-    }
-
-    pub fn get_num_members(&self, slug: String) -> usize {
-        let guild: Guild = self.guilds.get(&slug).unwrap_or_default().into();
-
-        guild.members.len()
-    }
-
-    /// Returns true if the guild exists
-    pub fn guild_exists(&self, slug: &String) ->bool {
-        match self.guilds.get(&slug) {
-            Some(value) => {
-                let log_message = format!("Guild exists {:?}", value);
-                env::log(log_message.as_bytes());
-                true
-            },
-            None => false
-        }
-    }
-
-    /// Returns true if the current account is member of the Guild
-    pub fn check_if_member(&self, slug: String) -> bool {
-        let guild: Guild;
-        match self.guild_exists(&slug) {
-            //true => let guild: Guild = self.guilds.get(&slug).expect("Guild does not exist").into();
-            true => guild = self.guilds.get(&slug).unwrap_or_default().into(),
-            false => {
-                let msg = format!("Guild {} does not exist", slug);
-                env::log(msg.as_bytes());
-                return false;
-            }
-        }
-
-        match guild.members.get(&env::predecessor_account_id()) {
-            Some(_member) => true,
-            None => false,
-        }
-    }
-
-    pub fn get_member_list(&self, slug: String) -> HashSet<AccountId> {
-        let guild: Guild;
-        match self.guild_exists(&slug) {
-            true => {
-                guild = self.guilds.get(&slug).unwrap_or_default().into();
-                return guild.members;
-            },
-            false => {
-                let msg = format!("Guild {} does not exist", slug);
-                env::log(msg.as_bytes());
-                return HashSet::<AccountId>::new();
-            }
-        };
-    }
-
-    //PROPOSAL: Methods for users and guild list by user
-    
-    //Checks if the current user is already registered in our platform
-    pub fn check_if_user(&self) -> bool {
-        match self.users.get(&env::predecessor_account_id()) {
-            Some(_user) => true,
-            None => {
-                let msg = format!("User {} is not part of any guild.", env::predecessor_account_id());
-                env::log(msg.as_bytes());
-                false
-            },
-        }
-    }
-
-    //Gets the list of guilds a user is part of
-    pub fn get_guilds_by_user(&self) -> HashSet<String> {
-        match self.check_if_user() {
-            true => {
-                self.users.get(&env::predecessor_account_id()).unwrap_or_default()
-            },
-            false => {
-                HashSet::<String>::new()
-            },
-        }
-    }
-}
-
-/// Returns an empty Guild
-/// TODO: Create a new "Constructor" trait?
-pub fn new_guild() -> Guild {
-    env::log(format!("Emplty guild created").as_bytes());
-    Guild {
-        slug: String::from(""),
-        title: String::from(""),
-        oneliner: String::from(""),
-        website: String::from(""),
-        app: String::from(""),
-        whitepaper: String::from(""),
-        twitter: String::from(""),
-        telegram: String::from(""),
-        discord: String::from(""),
-        medium: String::from(""),
-        github: String::from(""),
-        ticker: String::from(""),
-        logo: String::from(""),
-        contract_str: String::from(""),
-        youtube: String::from(""),
-        members: HashSet::new(),
-    }
-}
-
